delete Typing.Utils
