+++ compiler/Plato/Typing/Env.hs
@@ -10,7 +10,6 @@ module Plato.Typing.Env (
         ConEnv,
         initConEnv,
         HasConEnv (..),
-        splitConstrTy,
         extendConEnv,
 ) where
 
@@ -22,6 +21,7 @@ import Plato.Common.Error
 import Plato.Common.Ident
 import Plato.Common.Location
 import Plato.Syntax.Typing
+import Plato.Syntax.Typing.Helper
 
 data Bind
         = ValBind Type
@@ -87,13 +87,6 @@ class HasConEnv a where
 instance HasConEnv ConEnv where
         getConEnv = id
         modifyConEnv = id
-
-splitConstrTy :: Rho -> ([Sigma], Tau)
-splitConstrTy = go []
-    where
-        go :: [Sigma] -> Rho -> ([Sigma], Tau)
-        go acc (ArrT sigma rho) = go (unLoc sigma : acc) (unLoc rho)
-        go acc tau = (reverse acc, tau)
 
 extendConEnv :: HasConEnv env => Ident -> [TyVar] -> [(Ident, LType)] -> env -> env
 extendConEnv id params constrs = modifyConEnv $ M.insert id (params, constrs')
